freespeakerbureau_scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.4
lxml==4.9.3
tqdm==4.66.1
python-dotenv==1.0.0
pymongo==4.6.0